third_party/WebKit/LayoutTests/http/tests/serviceworker/onactivate-script-error.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharness-helpers.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
function wait_for_install(worker) {
  return new Promise(function(resolve, reject) {
      worker.addEventListener('statechange', function(event) {
          if (worker.state == 'installed')
            resolve();
          else if (worker.state == 'redundant')
            reject();
        });
    });
}

function wait_for_activate_event(worker) {
  return new Promise(function(resolve) {
      worker.addEventListener('statechange', function(event) {
          if (worker.state == 'activated')
            resolve(true);
          else if (worker.state == 'redundant')
            resolve(false);
        });
    });
}

function make_test(name, script, expect_activate) {
  promise_test(function(t) {
      var scope = script;
      var registration;
      return service_worker_unregister_and_register(t, script, scope)
        .then(function(r) {
            registration = r;
            return wait_for_install(registration.installing);
          })
        .then(function() {
            return wait_for_activate_event(registration.waiting);
          })
        .then(function(did_activate) {
            assert_equals(did_activate, expect_activate,
                          'The worker was activated');
          });
    }, name);
}

[
  {
    name: 'activate handler throws an error',
    script: 'resources/onactivate-throw-error-worker.js',
    expect_activate: false
  },
  {
    name: 'activate handler throws an error, error handler does not cancel',
    script: 'resources/onactivate-throw-error-with-empty-onerror-worker.js',
    expect_activate: false
  },
  {
    name: 'activate handler dispatches an event that throws an error',
    script: 'resources/onactivate-throw-error-from-nested-event-worker.js',
    expect_activate: true
  },
  {
    name: 'activate handler throws an error that is cancelled',
    script: 'resources/onactivate-throw-error-then-cancel-worker.js',
    expect_activate: true
  },
  {
    name: 'activate handler throws an error and prevents default',
    script: 'resources/onactivate-throw-error-then-prevent-default-worker.js',
    expect_activate: true
  }
].forEach(function(test_case) {
    make_test(test_case.name, test_case.script, test_case.expect_activate);
  });
</script>
